+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This is a list of manually mantained changes and updates for each version.
 
 ## Changed
 
-None
+### Router error for missing code will be status code
 
 ## Deprecated